requirements.txt
africastalking==1.2.5
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contextlib2==21.6.0
coverage==7.3.1
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.5
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schema==0.7.5
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
